# [HfFileSystem] Faster `fs.walk()`