...by using `expand_info=False` by default (same logic as `fs.glob()`)

before:

```python
In [1]: from huggingface_hub import HfFileSystem

In [2]: %time _ = list(HfFileSystem().walk("hf://datasets/allenai/c4/en"))
CPU times: user 275 ms, sys: 27.6 ms, total: 302 ms
Wall time: 11.6 s
```

after:

```python
In [1]: from huggingface_hub import HfFileSystem

In [2]: %time _ = list(HfFileSystem().walk("hf://datasets/allenai/c4/en"))
CPU times: user 176 ms, sys: 22.4 ms, total: 198 ms
Wall time: 3.25 s
```